pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "flask>=3.1.0",
    "nltk>=3.9.1",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "scikit-learn>=1.6.1",
    "spacy>=3.8.5",
]
